+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/TrackComponentFactory.java
@@ -18,20 +18,23 @@
  */
 package is.codion.framework.demos.chinook.ui;
 
+import is.codion.framework.demos.chinook.domain.Chinook.InvoiceLine;
 import is.codion.framework.domain.entity.Entity;
-import is.codion.framework.domain.entity.attribute.ForeignKey;
 import is.codion.swing.common.ui.component.value.ComponentValue;
 import is.codion.swing.framework.model.SwingEntityEditModel;
 import is.codion.swing.framework.ui.component.DefaultEntityComponentFactory;
 import is.codion.swing.framework.ui.component.EntitySearchField;
 
-public final class TrackComponentFactory extends DefaultEntityComponentFactory<Entity, ForeignKey, EntitySearchField> {
+final class TrackComponentFactory extends DefaultEntityComponentFactory<Entity, EntitySearchField> {
+
+	TrackComponentFactory() {
+		super(InvoiceLine.TRACK_FK);
+	}
 
 	@Override
-	public ComponentValue<Entity, EntitySearchField> componentValue(ForeignKey foreignKey,
-																																	SwingEntityEditModel editModel,
+	public ComponentValue<Entity, EntitySearchField> componentValue(SwingEntityEditModel editModel,
 																																	Entity initialValue) {
-		ComponentValue<Entity, EntitySearchField> componentValue = super.componentValue(foreignKey, editModel, initialValue);
+		ComponentValue<Entity, EntitySearchField> componentValue = super.componentValue(editModel, initialValue);
 		EntitySearchField trackSearchField = componentValue.component();
 		trackSearchField.selectorFactory().set(new TrackSelectorFactory());